+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/SoraMediaChannel.kt
@@ -804,8 +804,6 @@ class SoraMediaChannel @JvmOverloads constructor(
         when (notification.eventType) {
             "connection.created", "connection.destroyed" -> {
                 val attendees = ChannelAttendeesCount(
-                    numberOfDownstreams = notification.numberOfDownstreamConnections ?: 0,
-                    numberOfUpstreams = notification.numberOfUpstreamConnections ?: 0,
                     numberOfSendrecvConnections = notification.numberOfSendrecvConnections!!,
                     numberOfSendonlyConnections = notification.numberOfSendonlyConnections!!,
                     numberOfRecvonlyConnections = notification.numberOfRecvonlyConnections!!,